pyproject.toml
[project]
name = "auth-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-yasg (>=1.21.15,<2.0.0)",
]

[tool.poetry]
packages = [{include = "auth_system", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
